+++ src/Controllers/racingCountInputController.js
@@ -1,4 +1,3 @@
-import { Console } from '@woowacourse/mission-utils';
 import validateRacingCount from '../Models/validations/racingCountValidator.js';
 
 const validateRacingCountNumberType = racingCountInput => {
@@ -9,4 +8,3 @@ const validateRacingCountNumberType = racingCountInput => {
     return false;
   }
 };
-
